Got rid of shared preferences
Using hiveDB instead of shared preferences for storing app states such as language ID and first launch status. Discovered a new bug - language is not updated in diseases library and its subsequent screens. When language is switched in diseases list screen, it is not updated in plants list screen when navigated back.